+++ NEWS.md
@@ -7,6 +7,7 @@
 * Change the use of to RCurl::getURL at geocode_ide_uy()
 * Try change tiles_ide_uy() from raster to terra, but at the crop step the ptr slot is lost.
 * Improvement of plot_geouy
+* Remove tiles_ide_uy() function, because IDEuy change require methods
 
 
 ## geouy v0.1.9 (2020-03-20)